requirements.txt
fastapi==0.103.2
uvicorn==0.23.2
ultralytics==8.0.196
pillow==10.0.1
python-multipart==0.0.6
torch==2.0.1